Update 28.xml
Correction of transliteration, and typos in Syriac and English.
Addition of missing Arabic and revision and addition of URIs to works as per description.
Revision and addition of URIs to other works, persons and place names in XML or Excel sheet.
Addition of footnote notes where needed.
Issue in Github on arithmetical figures and Coptic characters, and slight issue on the disorder of Syriac display in one sentence (where the arithmetical figures appear in the element collation).