Update 437.xml
All planned changes in the description have been carried out apart from the last change on Missing various unspecified things, which are characters that cannot be displayed (Issue on Srophe).

Addition of bibliographic citations.
Correction of transliteration and typos,.

Correction of some of the ms parts and folios and <ref target of some of the ms parts (see Issue in Srophe).