Update 69.xml
Correction of transliteration and typos in Syriac and English, plus addition of locus for cited works within addition#3.
Addition of bibliographic citations.
Deletion of an empty element.
Some arithmetical figures could not be read where the Syriac is displayed incorrectly in the <physDesc>.
See issue on Srophe and on a long paragraph that might need to be divided.